docs: moved predicate deployment docs snippets
