+++ apps/docs-snippets2/sway/Forc.toml
@@ -1,6 +1,7 @@
 [workspace]
 members = [
   "bytecode-input",
+  "configurable-pin",
   "counter",
   "echo-asset-id",
   "echo-bytes",
