+++ apps/docs-snippets2/sway/configurable-pin/src/main.sw
@@ -1,0 +1,12 @@
+// #region predicate-with-configurable-pin-1
+predicate;
+
+configurable {
+    PIN: u64 = 1337,
+}
+
+fn main(pin: u64) -> bool {
+    return PIN == pin;
+}
+
+// #endregion predicate-with-configurable-pin-1
